Moved OracleSphereRenderer to mmstd_moldyn
