+++ core/src/factories/ModuleClassRegistry.cpp
@@ -28,7 +28,6 @@
 #include "mmcore/moldyn/MipDepthSphereRenderer.h"
 #include "mmcore/moldyn/MMPLDDataSource.h"
 #include "mmcore/moldyn/MMPLDWriter.h"
-#include "mmcore/moldyn/OracleSphereRenderer.h"
 #include "mmcore/moldyn/SimpleSphereRenderer.h"
 #include "mmcore/moldyn/ClusteredSphereRenderer.h"
 #include "mmcore/moldyn/DirPartColModulate.h"
@@ -95,7 +94,6 @@ void factories::register_module_classes(factories::ModuleDescriptionManager& ins
     instance.RegisterAutoDescription<moldyn::MMPLDWriter>();
     instance.RegisterAutoDescription<moldyn::SimpleSphereRenderer>();
     instance.RegisterAutoDescription<moldyn::ClusteredSphereRenderer>();
-    instance.RegisterAutoDescription<moldyn::OracleSphereRenderer>();
     instance.RegisterAutoDescription<moldyn::DirPartColModulate>();
     instance.RegisterAutoDescription<moldyn::ParticleListFilter>();
     instance.RegisterAutoDescription<moldyn::DirPartFilter>();
