+++ plugins/mmstd_moldyn/src/mmstd_moldyn.cpp
@@ -43,6 +43,7 @@
 #include "rendering/SimpleGeoSphereRenderer.h"
 #include "rendering/ArrowRenderer.h"
 #include "rendering/GrimRenderer.h"
+#include "rendering/OracleSphereRenderer.h"
 
 #include "mmstd_moldyn/EllipsoidalDataCall.h"
 #include "mmstd_moldyn/ParticleGridDataCall.h"
@@ -150,6 +151,7 @@ namespace {
             this->module_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::misc::DataGridder>();
             this->module_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::io::MMPGDWriter>();
             this->module_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::io::MMPGDDataSource>();
+            this->module_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::rendering::OracleSphereRenderer>();
             // register calls here:
             this->call_descriptions.RegisterAutoDescription< ::megamol::stdplugin::moldyn::BrickStatsCall>();
             this->call_descriptions.RegisterAutoDescription<megamol::stdplugin::moldyn::EllipsoidalParticleDataCall>();
